operators::common — types and methods generally useful for differential computation
Extracts from int_proxy the pieces whose correctness is a property of their
own inputs and outputs, so they can be used and tested in isolation:

  * TimeHistory — replay times ascending with meet-of-remainder tracking and a
    meet-advanced deduplicated buffer (a cheaper ValueHistory when only time
    structure matters). ValueHistory itself was already shared (operators::)
    and is re-exported here.
  * discover_times — the times at which a key's reduction must be
    re-evaluated, plus those to carry past 'upper'; O(times) peak memory.
    Its one cross-call obligation (seeds from the novel batch's own support)
    is stated at the signature.
  * bilinear_wave — the join of two histories in time order, work bounded by
    netted accumulation sizes rather than history lengths.
  * tile_descriptions — cut an output interval into batch descriptions along
    held capabilities.

int_proxy's tactics become expressions over these helpers, behaviorally
identical (its tests and oracles carry the equivalence). Written to host a
second tactic binding whose times are ranks into a caller-supplied store.
