+++ differential-dataflow/src/operators/mod.rs
@@ -9,6 +9,7 @@ pub use self::count::CountTotal;
 pub use self::threshold::ThresholdTotal;
 
 pub mod arrange;
+pub mod common;
 pub mod int_proxy;
 pub mod reduce;
 pub mod iterate;
@@ -88,21 +89,28 @@ impl<V: Copy, T: Ord + Lattice, D: crate::difference::Semigroup> EditList<V, T, 
     }
 }
 
-struct ValueHistory<V, T, D> {
+/// A loaded, time-ordered replay of one key's `(value, time, diff)` edits, with meet-advanced
+/// buffer collapse — the shared machinery under the cursor reduce, the `int_proxy` tactics, and
+/// the [`common`](crate::operators::common) helpers. Its local contract: after
+/// `advance_buffer_by(meet)` with the meet of the un-replayed times, the buffer is consolidated
+/// and replay cost stays linear in edits rather than quadratic.
+pub struct ValueHistory<V, T, D> {
     edits: EditList<V, T, D>,
     history: Vec<(T, T, usize, usize)>,     // (time, meet, value_index, edit_offset)
     buffer: Vec<((V, T), D)>,               // where we accumulate / collapse updates.
 }
 
 impl<V: Copy + Ord, T: Ord + Clone + Lattice, D: crate::difference::Semigroup> ValueHistory<V, T, D> {
-    fn new() -> Self {
+    /// An empty history, to be `load`ed.
+    pub fn new() -> Self {
         ValueHistory {
             edits: EditList::new(),
             history: Vec::new(),
             buffer: Vec::new(),
         }
     }
-    fn clear(&mut self) {
+    /// Discards all loaded state (capacity retained).
+    pub fn clear(&mut self) {
         self.edits.clear();
         self.history.clear();
         self.buffer.clear();
@@ -172,7 +180,7 @@ impl<V: Copy + Ord, T: Ord + Clone + Lattice, D: crate::difference::Semigroup> V
     /// cursor-free ingestion path: the `int_proxy` tactics present `(value_id, time, diff)`
     /// runs directly rather than through a `Cursor`, and share this machinery instead of
     /// re-implementing it. Ungrouped input is still correct, only less compact.
-    fn load_iter(&mut self, edits: impl Iterator<Item = (V, T, D)>, advance_by: Option<&T>) {
+    pub fn load_iter(&mut self, edits: impl Iterator<Item = (V, T, D)>, advance_by: Option<&T>) {
         self.edits.clear();
         let mut cur: Option<V> = None;
         for (v, mut time, diff) in edits {
@@ -190,37 +198,38 @@ impl<V: Copy + Ord, T: Ord + Clone + Lattice, D: crate::difference::Semigroup> V
 
 impl<V: Copy + Ord, T: Ord + Clone + Lattice, D: Clone + crate::difference::Semigroup> ValueHistory<V, T, D> {
     /// The next (least) un-replayed time.
-    fn time(&self) -> Option<&T> { self.history.last().map(|x| &x.0) }
+    pub fn time(&self) -> Option<&T> { self.history.last().map(|x| &x.0) }
     /// The meet of all un-replayed times.
-    fn meet(&self) -> Option<&T> { self.history.last().map(|x| &x.1) }
+    pub fn meet(&self) -> Option<&T> { self.history.last().map(|x| &x.1) }
     /// The next un-replayed edit, as `(value, time, diff)`.
-    fn edit(&self) -> Option<(V, &T, &D)> {
+    pub fn edit(&self) -> Option<(V, &T, &D)> {
         self.history.last().map(|&(ref t, _, v, e)| (self.edits.values[v].0, t, &self.edits.edits[e].1))
     }
     /// The buffered (stepped-in, advanced, consolidated) edits.
-    fn buffer(&self) -> &[((V, T), D)] { &self.buffer[..] }
+    pub fn buffer(&self) -> &[((V, T), D)] { &self.buffer[..] }
     /// Move the next edit into the buffer.
-    fn step(&mut self) {
+    pub fn step(&mut self) {
         let (time, _, value_index, edit_offset) = self.history.pop().unwrap();
         self.buffer.push(((self.edits.values[value_index].0, time), self.edits.edits[edit_offset].1.clone()));
     }
     /// Step edits while the next time equals `time`; true iff any did.
-    fn step_while_time_is(&mut self, time: &T) -> bool {
+    pub fn step_while_time_is(&mut self, time: &T) -> bool {
         let mut found = false;
         while self.time() == Some(time) { found = true; self.step(); }
         found
     }
     /// Step edits while the next time is `<= time` in the TOTAL order (a superset of the
     /// partially-ordered downset; readers filter the buffer by `less_equal` themselves).
-    fn step_through(&mut self, time: &T) {
+    pub fn step_through(&mut self, time: &T) {
         while self.time().is_some_and(|t| t <= time) { self.step(); }
     }
     /// Advance buffered times by `meet` and consolidate — the collapse that keeps replay linear.
-    fn advance_buffer_by(&mut self, meet: &T) {
+    pub fn advance_buffer_by(&mut self, meet: &T) {
         for element in self.buffer.iter_mut() { (element.0).1.join_assign(meet); }
         crate::consolidation::consolidate(&mut self.buffer);
     }
-    fn is_done(&self) -> bool { self.history.is_empty() }
+    /// True when every edit has been replayed.
+    pub fn is_done(&self) -> bool { self.history.is_empty() }
 }
 
 struct HistoryReplay<'history, V, T, D> {
